updating equilibrium sampling script
