+++ scripts/Generate_equilibrium_sampling.py
@@ -10,10 +10,11 @@ import matplotlib.pyplot as plt
 import pickle
 import torchani
 import torch
-from neutromeratio.constants import device, platform, kT, exclude_set_ANI, mols_with_charge
+from neutromeratio.constants import _get_names, device, platform, kT, exclude_set_ANI, mols_with_charge, multiple_stereobonds
 import sys
 import os
-
+from neutromeratio.analysis import setup_alchemical_system_and_energy_function
+from neutromeratio.ani import AlchemicalANI1ccx, AlchemicalANI1x, AlchemicalANI2x
 
 # name of the system
 idx = int(sys.argv[1])
@@ -23,16 +24,18 @@ n_steps = int(sys.argv[2])
 base_path = str(sys.argv[3])
 # env
 env = str(sys.argv[4])
-assert(env == 'droplet' or env == 'vacuum')
+
+if not (env == 'droplet' or env == 'vacuum'):
+    raise RuntimeError('Env must be `droplet` or `vacuum`. Aborting.')
 # diameter
 if env == 'droplet':
     diameter_in_angstrom = int(sys.argv[5])
+    if not diameter_in_angstrom or diameter_in_angstrom < 1:
+        raise RuntimeError('Diameter must be above 1 Angstrom')
 
 protocol = []
-exp_results = pickle.load(open('../data/exp_results.pickle', 'rb'))
-for name in sorted(exp_results):
-    if name in exclude_set_ANI + mols_with_charge:
-        continue
+names = neutromeratio.parameter_gradients._get_names()
+for name in names:
     for lamb in np.linspace(0, 1, 11):
         protocol.append((name, np.round(lamb, 2)))
 
@@ -40,76 +43,17 @@ name, lambda_value = protocol[idx-1]
 print(name)
 print(lambda_value)
 
+energy_function, tautomer, flipped = setup_alchemical_system_and_energy_function(
+    name=name,
+    ANImodel=AlchemicalANI1ccx,
+    env=env,
+    diameter=diameter_in_angstrom
+    base_path=base_path)
 
-t1_smiles = exp_results[name]['t1-smiles']
-t2_smiles = exp_results[name]['t2-smiles']
-
-os.makedirs(f"{base_path}/{name}", exist_ok=True)
-
-t_type, tautomers, flipped = neutromeratio.utils.generate_tautomer_class_stereobond_aware(name, t1_smiles, t2_smiles)
-tautomer = tautomers[0]
-tautomer.perform_tautomer_transformation()
-
-if env == 'droplet':
-    m = tautomer.add_droplet(tautomer.hybrid_topology, 
-                        tautomer.hybrid_coords, 
-                        diameter=diameter_in_angstrom * unit.angstrom,
-                        restrain_hydrogen_bonds=True,
-                        restrain_hydrogen_angles=False,
-                        top_file=f"{base_path}/{name}/{name}_in_droplet.pdb")
-else:
-    pdb_filepath = f"{base_path}/{name}/{name}.pdb"
-    try:
-        traj = md.load(pdb_filepath)
-    except OSError:
-        coordinates = tautomer.hybrid_coords
-        traj = md.Trajectory(coordinates.value_in_unit(unit.nanometer), tautomer.hybrid_topology)
-        traj.save_pdb(pdb_filepath)
-    # set coordinates #NOTE: note the xyz[0]
-    tautomer.hybrid_coords = traj.xyz[0] * unit.nanometer
-
-# define the alchemical atoms
-alchemical_atoms = [tautomer.hybrid_hydrogen_idx_at_lambda_1, tautomer.hybrid_hydrogen_idx_at_lambda_0]
-
-# set the ANI model
-model = neutromeratio.ani.LinearAlchemicalSingleTopologyANI(alchemical_atoms=alchemical_atoms)
-model = model.to(device)
-torch.set_num_threads(1)
-
-# perform initial sampling
-if env == 'vacuum':
-    energy_function = neutromeratio.ANI1_force_and_energy(
-        model=model,
-        atoms=tautomer.hybrid_atoms,
-        mol=None,
-    )
-else:
-    energy_function = neutromeratio.ANI1_force_and_energy(
-        model=model,
-        atoms=tautomer.ligand_in_water_atoms,
-        mol=None,
-    )
-
-
-for r in tautomer.ligand_restraints:
-    energy_function.add_restraint_to_lambda_protocol(r)
-
-for r in tautomer.hybrid_ligand_restraints:
-    energy_function.add_restraint_to_lambda_protocol(r)
+energy_and_force = lambda x: energy_function.calculate_force(x, lambda_value)
 
 if env == 'droplet':
-    tautomer.add_COM_for_hybrid_ligand(np.array([diameter_in_angstrom/2, diameter_in_angstrom/2, diameter_in_angstrom/2]) * unit.angstrom)
-
-    for r in tautomer.solvent_restraints:
-        energy_function.add_restraint_to_lambda_protocol(r)
-
-    for r in tautomer.com_restraints:
-        energy_function.add_restraint_to_lambda_protocol(r)
-
-
-energy_and_force = lambda x : energy_function.calculate_force(x, lambda_value)
-if env == 'droplet':
-    x0 = tautomer.ligand_in_water_coordinates
+    x0 = tautomer.get_lig
     langevin = neutromeratio.LangevinDynamics(atoms=tautomer.ligand_in_water_atoms,
                                         energy_and_force=energy_and_force)
 elif env == 'vacuum':
@@ -118,8 +62,8 @@ elif env == 'vacuum':
                                         energy_and_force=energy_and_force)
 else:
     raise RuntimeError()
-x0, e_history = energy_function.minimize(x0, maxiter=5000, lambda_value=lambda_value)
 
+x0, e_history = energy_function.minimize(x0, maxiter=5000, lambda_value=lambda_value)
 equilibrium_samples, energies, restraint_bias, stddev, ensemble_bias = langevin.run_dynamics(x0,
                                                                     n_steps=n_steps,
                                                                     stepsize=0.5*unit.femtosecond,
